+++ src_c/scrap_win.c
@@ -73,10 +73,10 @@ _convert_internal_type(char *type)
  * \param size The size of the buffer.
  * \return The length of the format name.
  */
-static int
+static size_t
 _lookup_clipboard_format(UINT format, char *buf, int size)
 {
-    int len;
+    size_t len;
     char *cpy;
 
     memset(buf, 0, size);
@@ -118,7 +118,7 @@ _lookup_clipboard_format(UINT format, char *buf, int size)
  * \return The character buffer containing the BMP information.
  */
 static char *
-_create_dib_buffer(char *data, unsigned long *count)
+_create_dib_buffer(char *data, size_t *count)
 {
     BITMAPFILEHEADER hdr;
     LPBITMAPINFOHEADER bihdr;
@@ -184,10 +184,10 @@ pygame_scrap_lost(void)
 }
 
 int
-pygame_scrap_put(char *type, int srclen, char *src)
+pygame_scrap_put(char *type, Py_ssize_t srclen, char *src)
 {
     UINT format;
-    int nulledlen = srclen + 1;
+    Py_ssize_t nulledlen = srclen + 1;
     HANDLE hMem;
 
     if (!pygame_scrap_initialized()) {
@@ -235,7 +235,7 @@ pygame_scrap_put(char *type, int srclen, char *src)
 }
 
 char *
-pygame_scrap_get(char *type, unsigned long *count)
+pygame_scrap_get(char *type, size_t *count)
 {
     UINT format = _convert_format(type);
     char *retval = NULL;
@@ -316,7 +316,7 @@ pygame_scrap_get_types(void)
     char **tmptypes;
     int i = 0;
     int count = -1;
-    int len;
+    size_t len;
     char tmp[100] = {'\0'};
     int size = 0;
